Implement context with exported composed types
This change allows for "easier" creation of implementations
of a Context type. Because it's composed of many different
exported types that are based on simple types. Out of 36
methods on the Context interface, 28 of them can be implemented
by simply embedding a few structs into your mock type.

These new types are:
- echo.RequestReader
- echo.ResponseWriter
- echo.MapStore
- echo.NetContextEmbedder

This means the only methods that are required for Context
implementation is:
- Path
- P
- Param
- ParamNames
- Bind
- Render
- File
- Error
- Echo
- Handler
- Logger
- ServeContent
- Reset

Changes to Context()'s public interface:
- Delete the Object() method from the interface
  which made it impossible to implement this interface since
  it returned an exported type.
- Add a Del() to be able to delete things stored in "store".