Aliased slices should be consistent with builtin slices In working with an API at work we had the use case on query params to accept the following syntaxes:
- `?foo=1&foo=2&foo=3`
- `?foo=1,2,3`

Initially we were using only the former syntax, but then a new library was
developed that sent the latter syntax.  When we switched from defining our field
as `[]int` to a custom type `Ints` that implemented `UnmarshalParam` we found
that when there were multiple values for the same key, the `UnmarshalParam`
would only receive the first value and others would be lost.

This change brings
consistency between the builtin slices and the aliased slices for query
parameters.